[I25-170] refactor : 주석 및 types 폴더 제거
