+++ src/utils/convertStudentNumber.ts
@@ -1,3 +1,4 @@
+// 0000 형식의 학번을 '0반 0번' 형식으로 바꿔주는 함수
 export const convertStudentNumber = (student_number: number | null): string => {
   if (!student_number) return '';
